requirements.txt
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
factory-boy==3.2.1
Faker==14.2.0
fastapi==0.79.0
flake8==5.0.4
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.7.0
orjson==3.7.8
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pydantic==1.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.11
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
